+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Unreleased
+
+- Fixed a bug when `server:grep_log()` failed to find a string logged in
+  `server:exec()` called immediately before it (gh-421).
+
 ## 1.1.0
 
 - Added logging to unified file (gh-324).